Fix reconnect command

+++ src/extension.ts
@@ -8,6 +8,8 @@ import { ObjectTreeDataProvider, ObjectTreeItem } from "./panels/ObjectExplorer"
 import { TagEditorPanel } from "./panels/TagEditorPanel"
 import { TagEntry } from "./types/EngineTag"
 
+let connected = false;
+
 export function activate(context: vscode.ExtensionContext) {
 	const rootPath =
 		vscode.workspace.workspaceFolders &&
@@ -15,14 +17,22 @@ export function activate(context: vscode.ExtensionContext) {
 			? vscode.workspace.workspaceFolders[0].uri.fsPath
 			: undefined
 
-	const connectDevkit = () => {
-		// Connect to devkit server
+	const connectDevkitServer = () => {
 		initializeDevkitClient(
 			vscode.workspace.getConfiguration("balltzeDevkit").get("host") || "localhost",
-			vscode.workspace.getConfiguration("balltzeDevkit").get("port") || 19190
+			vscode.workspace.getConfiguration("balltzeDevkit").get("port") || 19190,
+			() => {
+				vscode.window.showInformationMessage("Connected to devkit server");
+				connected = true;
+			},
+			() => {
+				vscode.window.showErrorMessage("Failed to connect to devkit server");
+				connected = false;
+			}
 		)
 	}
-	connectDevkit()
+
+	connectDevkitServer();
 
 	const tagTreeProvider = new TagsTreeDataProvider(rootPath)
 	const tagsExplorer = vscode.window.createTreeView("tagsExplorer", {
@@ -36,61 +46,78 @@ export function activate(context: vscode.ExtensionContext) {
 		{
 			command: "balltzeDevkit.reconnect",
 			action: () => {
-				connectDevkit()
-				tagTreeProvider.refresh()
-				objectTreeProvider.refresh()
+				connectDevkitServer();
+				tagTreeProvider.refresh();
+				objectTreeProvider.refresh();
 			}
 		},
 		{
 			command: "tagsExplorer.refresh",
-			action: () => tagTreeProvider.refresh()
+			action: () => {
+				if(connected) {
+					tagTreeProvider.refresh();
+				}
+				else {
+					vscode.window.showErrorMessage("Not connected to devkit server");
+				}
+			}
 		},
 		{
 			command: "objectsExplorer.refresh",
-			action: () => objectTreeProvider.refresh()
+			action: () => {
+				if(connected) {
+					objectTreeProvider.refresh();
+				}
+				else {
+					vscode.window.showErrorMessage("Not connected to devkit server");
+				}
+			}
 		},
 		{
 			command: "tagsExplorer.copyPath",
 			action: (item: TagTreeItem) => {
-				// Remove the leading slash
-				vscode.env.clipboard.writeText(item.path.slice(1))
+				vscode.env.clipboard.writeText(item.path.slice(1));
 			}
 		},
 		{
 			command: "tagsExplorer.copyTagHandle",
 			action: (item: TagTreeItem) => {
-				vscode.env.clipboard.writeText(item.tagEntry?.handle.toString() || "")
+				vscode.env.clipboard.writeText(item.tagEntry?.handle.toString() || "");
 			}
 		},
 		{
 			command: "tagsExplorer.spawn",
 			action: async (item: TagTreeItem) => {
-				await tagTreeProvider.spawnObject(item)
+				if(connected) {
+					if(item.tagEntry) {
+						await tagTreeProvider.spawnObject(item);
+						vscode.window.showInformationMessage(`Spawned ${item.label}`);
+					}
+					else {
+						vscode.window.showErrorMessage("Failed to spawn tag: tag entry not found");
+					}
+				}
+				else {
+					vscode.window.showErrorMessage("Not connected to devkit server");
+				}
 			}
 		},
 		{
 			command: "objectsExplorer.delete",
 			action: (item: ObjectTreeItem) => {
-				vscode.window.showInformationMessage(`Deleting ${item.label}`)
-				objectTreeProvider.deleteItem(item)
+				if(connected) {
+					objectTreeProvider.deleteItem(item);
+					vscode.window.showInformationMessage(`Deleted ${item.label}`);
+				}
+				else {
+					vscode.window.showErrorMessage("Not connected to devkit server");
+				}
 			}
 		},
 		{
 			command: "tagsExplorer.openEditor",
 			action: (node: TagTreeItem) => {
-				if(node.isFile && node.tagEntry) {
-					TagEditorPanel.render(context, node.tagEntry);
-				}
-				else {
-					vscode.window.showErrorMessage(`Failed to open tag editor panel: ${node.path} is not a file`);
-				}
-			}
-		},
-		{
-			command: "tagsExplorer.openEditorByHandle",
-			action: (handle: number) => {
-				let node = tagTreeProvider.getTreeItemByTagHandle(handle);
-				if(node) {
+				if(connected) {
 					if(node.isFile && node.tagEntry) {
 						TagEditorPanel.render(context, node.tagEntry);
 					}
@@ -99,7 +126,29 @@ export function activate(context: vscode.ExtensionContext) {
 					}
 				}
 				else {
-					vscode.window.showErrorMessage(`Failed to open tag editor panel for handle ${handle}`);
+					vscode.window.showErrorMessage("Not connected to devkit server");
+				}
+			}
+		},
+		{
+			command: "tagsExplorer.openEditorByHandle",
+			action: (handle: number) => {
+				if(connected) {
+					let node = tagTreeProvider.getTreeItemByTagHandle(handle);
+					if(node) {
+						if(node.isFile && node.tagEntry) {
+							TagEditorPanel.render(context, node.tagEntry);
+						}
+						else {
+							vscode.window.showErrorMessage(`Failed to open tag editor panel: ${node.path} is not a file`);
+						}
+					}
+					else {
+						vscode.window.showErrorMessage(`Failed to open tag editor panel for handle ${handle}`);
+					}
+				}
+				else {
+					vscode.window.showErrorMessage("Not connected to devkit server");
 				}
 			}
 		},
@@ -130,6 +179,6 @@ export function activate(context: vscode.ExtensionContext) {
 	]
 
 	commands.forEach(({ command, action }) => {
-		vscode.commands.registerCommand(command, action)
+		vscode.commands.registerCommand(command, action);
 	});
 }
